Поправлен вывод ошибки на agent-е
Из документации по zerolog https://github.com/rs/zerolog#leveled-logging:
```
It is very important to note that when using the zerolog chaining API, as shown above (log.Info().Msg("hello world"), the chain must have either the Msg or Msgf method call. If you forget to add either of these, the log will not occur and there is no compile time error to alert you of this.
```